# backend/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "AnnouncementCreate": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "pattern": "^\\d+\\.\\d+\\.\\d+(?:-[0-9A-Za-z.-]+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "audience": {
            "default": "all",
            "enum": [
              "all",
              "consumer",
              "business"
            ],
            "title": "Audience",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "enum": [
                  "system",
                  "update",
                  "activity",
                  "maintenance"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "content": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "presentation": {
            "anyOf": [
              {
                "enum": [
                  "banner",
                  "modal"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presentation"
          },
          "priority": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "severity": {
            "default": "info",
            "enum": [
              "info",
              "important",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "starts_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "published",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "AnnouncementCreate",
        "type": "object"
      },
      "AnnouncementUpdate": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "pattern": "^\\d+\\.\\d+\\.\\d+(?:-[0-9A-Za-z.-]+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "audience": {
            "anyOf": [
              {
                "enum": [
                  "all",
                  "consumer",
                  "business"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience"
          },
          "category": {
            "anyOf": [
              {
                "enum": [
                  "system",
                  "update",
                  "activity",
                  "maintenance"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "content": {
            "anyOf": [
              {
                "maxLength": 10000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "presentation": {
            "anyOf": [
              {
                "enum": [
                  "banner",
                  "modal"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presentation"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "severity": {
            "anyOf": [
              {
                "enum": [
                  "info",
                  "important",
                  "critical"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "starts_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "published",
                  "expired"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "AnnouncementUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "跨境电商赛训效率工具箱",
    "version": "1.8.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/announcements": {
      "get": {
        "operationId": "list_announcements_api_announcements_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Announcements Api Announcements Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Announcements",
        "tags": [
          "公告管理"
        ]
      },
      "post": {
        "operationId": "create_announcement_api_announcements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Announcement Api Announcements Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Announcement",
        "tags": [
          "公告管理"
        ]
      }
    },
...
    "/api/announcements/feed": {
      "get": {
        "operationId": "get_announcement_feed_api_announcements_feed_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Announcement Feed Api Announcements Feed Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Announcement Feed",
        "tags": [
          "公告管理"
        ]
      }
    },
    "/api/announcements/release-notes/{version}": {
      "get": {
        "operationId": "get_release_notes_api_announcements_release_notes__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Release Notes Api Announcements Release Notes  Version  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Release Notes",
        "tags": [
          "公告管理"
        ]
      }
    },
    "/api/announcements/{announcement_id}": {
      "delete": {
        "operationId": "delete_announcement_api_announcements__announcement_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "title": "Announcement Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Delete Announcement Api Announcements  Announcement Id  Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Announcement",
        "tags": [
          "公告管理"
        ]
      },
      "put": {
        "operationId": "update_announcement_api_announcements__announcement_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "title": "Announcement Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Announcement Api Announcements  Announcement Id  Put"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Announcement",
        "tags": [
          "公告管理"
        ]
      }
    },
    "/api/announcements/{announcement_id}/dismiss": {
      "post": {
        "operationId": "dismiss_announcement_api_announcements__announcement_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "title": "Announcement Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Dismiss Announcement Api Announcements  Announcement Id  Dismiss Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dismiss Announcement",
        "tags": [
          "公告管理"
        ]
      }
    },
    "/api/announcements/{announcement_id}/read": {
      "post": {
        "operationId": "mark_announcement_read_api_announcements__announcement_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "title": "Announcement Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Mark Announcement Read Api Announcements  Announcement Id  Read Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Announcement Read",
        "tags": [
          "公告管理"
        ]
      }
    },
...
  }
}